requirements.txt
openai
langchain
langchain_community
gradio 
pypdf
weaviate-client 
sentence-transformers  
PyPDF2